ref response model; techTaskmetr

+++ techTaskmetr/internal/models/model.go
@@ -8,5 +8,5 @@ type Task struct {
 	Description string    `gorm:"description" json:"description"`
 	CreatedAt   time.Time `gorm:"created_at" json:"created_at"`
 	Status      string    `gorm:"status" json:"status"`
-	UserID      int64     `gorm:"user_id" json:"user_id"`
+	UserID      int       `gorm:"user_id" json:"user_id"`
 }
